Number Pad + LCD + Pump Re-configure
Code provided adds a number pad and LCD (16x2) screen to operate in a similar fashion to a syringe stepper motor pump. The addition of a screen and number pad allows the system to be entirely portable and self-standing without the need for a computer. The code provided retains the usage via computer. The On-board code should have pre-sets for operation and a basic flow chart will be provided on the usage of the device and how flow rate is calculated based on syringe diameter, etc.

Currently this is beta - Some bugs are present but it does somewhat operate as expected. Known issues are input of certain values due to the in beta unification of number entry.