Beta version of DCPumpController with Numpad & LCD
Initial beta version of a standalone DCPumpController with Numpad. Does have glitches. The aim is to create a standalone DCPumpController with Numpad and 16x2 LCD screen to emulate a stepper motor syringe pump. The setup will work with the standard Lego syringe pump setup but it is designed for more intricate gear ratio setup.